requirements.txt
Flask
twilio
google-generativeai
requests
python-dotenv
gunicorn
firebase-admin
langdetect
python-dateutil